refactor: 아이템 구현 (#9)
